v0.4.5 Phase 0: Risky Mode rebaseline + scalping framing + secure OpenAI key UX
Operator directives 2026-05-17 — three corrections folded into one
ship-gate-friendly commit:

§7.2 — f64 everywhere except GPU (was f32-throughout-Risky-Mode):
  - crates/forex-core/src/domain/risky_mode.rs rewritten from scratch:
    bankroll/PnL/risk-fraction numerics all f64, doc updated.
  - RiskyModeManager::check_trade_allowed, calculate_position_size_usd,
    record_trade_outcome and the snapshot fields on TradingSession's
    RiskyModeState all f64.
  - Auto-trade signal confidence + AUTO_TRADE_MIN_CONFIDENCE +
    BotDecisionEntry.confidence: f32 -> f64.
  - orders.rs Risky Mode gate composition: size_usd / sl_pips /
    tp_pips passed as f64; close-path record_trade_outcome no longer
    casts realized_pnl through f32.
  - WizardConfig.risky_mode_ruin_ceiling_acknowledged: f32 -> f64.

§7.1 — Risky Mode is autonomous-only, 30-50% per trade, 99% S1 ruin
acceptance, NOT FTMO-related:
  - New RiskyModeConfig.autonomous_only_contract_accepted: bool —
    RiskyModeManager::new() now REJECTS construction without it
    (informed-consent gate on the operator).
  - New KillSwitchTier::ManualOrderWhileAutonomousOnly variant +
    RiskyModeManager::rejects_manual_orders() — Phase 0d will wire
    OrderSource into execute_ctrader_order so the manual-order path
    rejects via this gate.
  - Per-trade sizing knob renamed kelly_fraction_multiplier ->
    risk_per_trade_fraction; validator enforces the operator-stated
    [0.30, 0.50] band per stage.
  - MAX_ACCEPTABLE_INITIAL_RUIN_PROBABILITY raised to 0.99.
  - autonomy_risk.rs wizard step: stale "Kelly-tapered, 50% ruin"
    framing replaced with the operator's "autonomous, 30-50%, 99%
    ruin, scalp many, target net profit after expenses" language.
    Stale #![allow(dead_code)] silencer dropped.
  - trading_tests.rs: signed_risky_mode_config() helper threads the
    contract through all 3 Risky-Mode integration test sites; f32
    test literals dropped.

Strategy framing fix (2026-05-17 directive — "δεν χρειαζεται 20 pips
την ημερα single trade, αφησε το bot να scalp πολλες φορες και να
παραγαγει pure profit μετα τα expenses"):
  - Retired the "20 pips per day, single high-conviction trade"
    framing in risky_mode.rs module doc.
  - Replaced with the scalp-many-times contract:
      * no hardcoded trades-per-day cap on the producer
      * optimisation target = net profit after commission + spread +
        swap, not a fixed pip count
      * DEFAULT_RISKY_TRADES_PER_DAY bumped 1.0 -> 10.0 (a throughput
        ASSUMPTION used only by the days-to-target estimator; has
        zero gating effect)
      * New RiskyModeConfig.expected_trades_per_day field — operator
        can tune per broker
  - Ruin-probability + days-to-target estimator now consult two new
    operator-configurable knobs:
      * RiskyModeConfig.expected_win_rate (default 0.52 — honest
        post-cost baseline)
      * RiskyModeConfig.expected_reward_to_risk (default 1.5)
    Previously hard-coded at (0.55, 3.0) which contradicted §7.1's
    99% ruin framing — at the previous defaults the model gave
    P(ruin) ~0.17, falsely optimistic. Honest defaults give
    P(ruin) ~1.0 at stage 0 (matching the framing the operator signed).

Hardcoded OpenAI key UX hazard removed:
  - crates/forex-news/src/openai.rs no longer reads OPENAI_API_KEY
    from env::var(). Constructor takes Option<SecretString> + model
    explicitly; an empty/None key disables the scorer (short-circuit
    with warn + neutral score). The wizard / settings UI is now the
    sole source of the operator's key.
  - Earlier behaviour silently picked up the dev's personal key on
    any machine with OPENAI_API_KEY exported — a quiet billing leak
    and a UX bug that the v0.4.5 ship-gate cannot carry.

Tests (all green):
  - forex-core/src/domain/risky_mode tests: 30 passed (added 4 new
    tests for cadence scaling, edge sensitivity, and validator
    rejections of the new knobs).
  - forex-news/src/openai tests: 7 passed (added 7 new tests for
    the constructor contract).
  - cargo check -p forex-core -p forex-app -p forex-news --tests:
    0 errors. 58 warnings in forex-app (separate cleanup — most are
    revertable C3-rework leftovers + a handful of deprecated egui
    calls).

DEFERRED to follow-up phases (clearly enumerated, not silent):
  - Phase 0c (next): cTrader wizard Step 4 must stop asking the user
    for client_id / client_secret (those are DEVELOPER credentials,
    not user credentials) and offer (a) embedded-app OAuth and
    (b) broker-account + investor/master-password for brokers that
    don't support OAuth.
  - Phase 0d (next-after-0c): wire OrderSource { Manual, Ai } through
    execute_ctrader_order so the manual-order path returns
    KillSwitchTier::ManualOrderWhileAutonomousOnly when
    rejects_manual_orders() is true.
  - Phase D2 (later): surface the new RiskyModeConfig knobs
    (expected_win_rate, expected_reward_to_risk, expected_trades_per_day)
    in the wizard's AutonomyRisk step + Settings panel.

+++ crates/forex-app/src/app_services/trading/auto_trade.rs
@@ -65,7 +65,7 @@ pub struct AutoTradeSignal {
     /// Confidence in `[0.0, 1.0]`. Compared against
     /// [`AUTO_TRADE_MIN_CONFIDENCE`] in the gate; signals below
     /// threshold are rejected with `GateDecision::BelowConfidence`.
-    pub confidence: f32,
+    pub confidence: f64,
     /// Display label rendered on the chart overlay marker after
     /// dispatch (`"AI BUY · 0.74"`, `"AI SELL · 0.81"`).
     pub label: String,
@@ -86,7 +86,7 @@ pub enum AutoTradeSide {
 /// Hard-coded at the spec floor — research §4.6.1 names 0.6 as the
 /// stage-0 minimum; the Risky Mode gate tightens this further per
 /// stage when active.
-pub const AUTO_TRADE_MIN_CONFIDENCE: f32 = 0.6;
+pub const AUTO_TRADE_MIN_CONFIDENCE: f64 = 0.6;
 
 /// Outcome of pushing an auto-trade signal through the gate chain.
 /// Every variant is observable from outside so the inference loop's
@@ -106,7 +106,7 @@ pub enum GateDecision {
     /// `side == AutoTradeSide::Flat` — no order to send.
     FlatSide,
     /// Confidence below [`AUTO_TRADE_MIN_CONFIDENCE`].
-    BelowConfidence { confidence: f32, minimum: f32 },
+    BelowConfidence { confidence: f64, minimum: f64 },
     /// News blackout active. Same gate that the manual path uses
     /// (`news_filter.is_blackout()`).
     NewsBlackout,
@@ -269,7 +269,7 @@ mod tests {
         s
     }
 
-    fn sample_signal(side: AutoTradeSide, confidence: f32) -> AutoTradeSignal {
+    fn sample_signal(side: AutoTradeSide, confidence: f64) -> AutoTradeSignal {
         AutoTradeSignal {
             symbol: "EURUSD".to_string(),
             side,
